The ship_hit() function has been changed. Now, when a ship collides with aliens, the fleet remains in the same position, taking into account the destroyed enemy ships. However, the player's ship changes and the player has fewer attempts left.
The check for the alien fleet reaching the bottom edge of the screen has been removed from update_aliens() and added to update_ships(). Now, when this event occurs, the player's ships change as well as during a collision.

A dictionary has been added to the GameStats class to store the sum of the points earned by the player at each level.

Fixed a bug with scoring after the fleet reaches the bottom edge of the screen. Now it is reset to the value obtained when moving to the current level.